perf(web): reduce gpu usage and split oversized backend files
Remove unnecessary will-change hints from hero title, sea battle animations,
cascade spin arrow, and pulse ring. pause game music equalizer canvas when
not playing. cap pixijs resolution at 2x and disable antialiasing in
glimworm game. extract bot targeting logic and checkers test helpers into
separate files to satisfy 500-line limit.
