+++ apps/web/src/app/[locale]/home/components/styles/home-bundle.scss
@@ -27,7 +27,6 @@
     /* Deep shadow */ 12px 18px 30px rgba(0, 0, 0, 0.6); /* Soft floor shadow */
 
   animation: hero-float-3d 6s ease-in-out infinite;
-  will-change: transform;
   transition: transform 0.2s cubic-bezier(0.175, 0.885, 0.32, 1.275);
 }
 
@@ -74,9 +73,6 @@
 
   /* Internal Glow Filter - Makes the rainbow POP off the dark base */
   filter: drop-shadow(0 0 1px rgba(255, 255, 255, 0.5));
-
-  /* Lifted physically and visually */
-  transform: translateZ(5px);
 
   animation:
     hero-shimmer 8s linear infinite,
