+++ apps/web/src/app/styles/animations.scss
@@ -339,7 +339,6 @@
 
 .pulse-ring {
   animation: playButtonPulse 3s infinite;
-  will-change: transform, opacity;
 }
 
 @keyframes scaleIn {
