+++ apps/web/src/widgets/CascadeGame/ui/CascadeGame.module.css
@@ -638,7 +638,6 @@
   display: inline-block;
   font-size: 16px;
   animation: cascade-spin 6s linear infinite;
-  will-change: transform;
 }
 .dirArrowReverse {
   animation-direction: reverse;
